+++ packages/nns/candid/governance.certified.idl.js
@@ -41,6 +41,7 @@ export const idlFactory = ({ IDL }) => {
   const IncreaseDissolveDelay = IDL.Record({
     'additional_dissolve_delay_seconds' : IDL.Nat32,
   });
+  const SetVisibility = IDL.Record({ 'visibility' : IDL.Opt(IDL.Int32) });
   const SetDissolveTimestamp = IDL.Record({
     'dissolve_timestamp_seconds' : IDL.Nat64,
   });
@@ -51,6 +52,7 @@ export const idlFactory = ({ IDL }) => {
     'StopDissolving' : IDL.Record({}),
     'StartDissolving' : IDL.Record({}),
     'IncreaseDissolveDelay' : IncreaseDissolveDelay,
+    'SetVisibility' : SetVisibility,
     'JoinCommunityFund' : IDL.Record({}),
     'LeaveCommunityFund' : IDL.Record({}),
     'SetDissolveTimestamp' : SetDissolveTimestamp,
@@ -101,12 +103,29 @@ export const idlFactory = ({ IDL }) => {
     'command' : IDL.Opt(Command),
     'neuron_id_or_subaccount' : IDL.Opt(NeuronIdOrSubaccount),
   });
+  const Controllers = IDL.Record({ 'controllers' : IDL.Vec(IDL.Principal) });
+  const CanisterSettings = IDL.Record({
+    'freezing_threshold' : IDL.Opt(IDL.Nat64),
+    'controllers' : IDL.Opt(Controllers),
+    'log_visibility' : IDL.Opt(IDL.Int32),
+    'wasm_memory_limit' : IDL.Opt(IDL.Nat64),
+    'memory_allocation' : IDL.Opt(IDL.Nat64),
+    'compute_allocation' : IDL.Opt(IDL.Nat64),
+  });
+  const UpdateCanisterSettings = IDL.Record({
+    'canister_id' : IDL.Opt(IDL.Principal),
+    'settings' : IDL.Opt(CanisterSettings),
+  });
   const InstallCode = IDL.Record({
     'arg' : IDL.Opt(IDL.Vec(IDL.Nat8)),
     'wasm_module' : IDL.Opt(IDL.Vec(IDL.Nat8)),
     'skip_stopping_before_installing' : IDL.Opt(IDL.Bool),
     'canister_id' : IDL.Opt(IDL.Principal),
     'install_mode' : IDL.Opt(IDL.Int32),
+  });
+  const StopOrStartCanister = IDL.Record({
+    'action' : IDL.Opt(IDL.Int32),
+    'canister_id' : IDL.Opt(IDL.Principal),
   });
   const Percentage = IDL.Record({ 'basis_points' : IDL.Opt(IDL.Nat64) });
   const Duration = IDL.Record({ 'seconds' : IDL.Opt(IDL.Nat64) });
@@ -288,7 +307,9 @@ export const idlFactory = ({ IDL }) => {
   const Action = IDL.Variant({
     'RegisterKnownNeuron' : KnownNeuron,
     'ManageNeuron' : ManageNeuron,
+    'UpdateCanisterSettings' : UpdateCanisterSettings,
     'InstallCode' : InstallCode,
+    'StopOrStartCanister' : StopOrStartCanister,
     'CreateServiceNervousSystem' : CreateServiceNervousSystem,
     'ExecuteNnsFunction' : ExecuteNnsFunction,
     'RewardNodeProvider' : RewardNodeProvider,
@@ -455,10 +476,12 @@ export const idlFactory = ({ IDL }) => {
   });
   const CfNeuron = IDL.Record({
     'has_created_neuron_recipes' : IDL.Opt(IDL.Bool),
+    'hotkeys' : IDL.Opt(Principals),
     'nns_neuron_id' : IDL.Nat64,
     'amount_icp_e8s' : IDL.Nat64,
   });
   const CfParticipant = IDL.Record({
+    'controller' : IDL.Opt(IDL.Principal),
     'hotkey_principal' : IDL.Text,
     'cf_neurons' : IDL.Vec(CfNeuron),
   });
@@ -600,6 +623,7 @@ export const idlFactory = ({ IDL }) => {
     'dissolve_state' : IDL.Opt(DissolveState),
     'followees' : IDL.Vec(IDL.Tuple(IDL.Int32, Followees)),
     'neuron_fees_e8s' : IDL.Nat64,
+    'visibility' : IDL.Opt(IDL.Int32),
     'transfer' : IDL.Opt(NeuronStakeTransfer),
     'known_neuron_data' : IDL.Opt(KnownNeuronData),
     'spawn_at_timestamp_seconds' : IDL.Opt(IDL.Nat64),
@@ -658,6 +682,7 @@ export const idlFactory = ({ IDL }) => {
     'stake_e8s' : IDL.Nat64,
     'joined_community_fund_timestamp_seconds' : IDL.Opt(IDL.Nat64),
     'retrieved_at_timestamp_seconds' : IDL.Nat64,
+    'visibility' : IDL.Opt(IDL.Int32),
     'known_neuron_data' : IDL.Opt(KnownNeuronData),
     'voting_power' : IDL.Nat64,
     'age_seconds' : IDL.Nat64,
@@ -924,6 +949,7 @@ export const init = ({ IDL }) => {
   const IncreaseDissolveDelay = IDL.Record({
     'additional_dissolve_delay_seconds' : IDL.Nat32,
   });
+  const SetVisibility = IDL.Record({ 'visibility' : IDL.Opt(IDL.Int32) });
   const SetDissolveTimestamp = IDL.Record({
     'dissolve_timestamp_seconds' : IDL.Nat64,
   });
@@ -934,6 +960,7 @@ export const init = ({ IDL }) => {
     'StopDissolving' : IDL.Record({}),
     'StartDissolving' : IDL.Record({}),
     'IncreaseDissolveDelay' : IncreaseDissolveDelay,
+    'SetVisibility' : SetVisibility,
     'JoinCommunityFund' : IDL.Record({}),
     'LeaveCommunityFund' : IDL.Record({}),
     'SetDissolveTimestamp' : SetDissolveTimestamp,
@@ -984,12 +1011,29 @@ export const init = ({ IDL }) => {
     'command' : IDL.Opt(Command),
     'neuron_id_or_subaccount' : IDL.Opt(NeuronIdOrSubaccount),
   });
+  const Controllers = IDL.Record({ 'controllers' : IDL.Vec(IDL.Principal) });
+  const CanisterSettings = IDL.Record({
+    'freezing_threshold' : IDL.Opt(IDL.Nat64),
+    'controllers' : IDL.Opt(Controllers),
+    'log_visibility' : IDL.Opt(IDL.Int32),
+    'wasm_memory_limit' : IDL.Opt(IDL.Nat64),
+    'memory_allocation' : IDL.Opt(IDL.Nat64),
+    'compute_allocation' : IDL.Opt(IDL.Nat64),
+  });
+  const UpdateCanisterSettings = IDL.Record({
+    'canister_id' : IDL.Opt(IDL.Principal),
+    'settings' : IDL.Opt(CanisterSettings),
+  });
   const InstallCode = IDL.Record({
     'arg' : IDL.Opt(IDL.Vec(IDL.Nat8)),
     'wasm_module' : IDL.Opt(IDL.Vec(IDL.Nat8)),
     'skip_stopping_before_installing' : IDL.Opt(IDL.Bool),
     'canister_id' : IDL.Opt(IDL.Principal),
     'install_mode' : IDL.Opt(IDL.Int32),
+  });
+  const StopOrStartCanister = IDL.Record({
+    'action' : IDL.Opt(IDL.Int32),
+    'canister_id' : IDL.Opt(IDL.Principal),
   });
   const Percentage = IDL.Record({ 'basis_points' : IDL.Opt(IDL.Nat64) });
   const Duration = IDL.Record({ 'seconds' : IDL.Opt(IDL.Nat64) });
@@ -1171,7 +1215,9 @@ export const init = ({ IDL }) => {
   const Action = IDL.Variant({
     'RegisterKnownNeuron' : KnownNeuron,
     'ManageNeuron' : ManageNeuron,
+    'UpdateCanisterSettings' : UpdateCanisterSettings,
     'InstallCode' : InstallCode,
+    'StopOrStartCanister' : StopOrStartCanister,
     'CreateServiceNervousSystem' : CreateServiceNervousSystem,
     'ExecuteNnsFunction' : ExecuteNnsFunction,
     'RewardNodeProvider' : RewardNodeProvider,
@@ -1338,10 +1384,12 @@ export const init = ({ IDL }) => {
   });
   const CfNeuron = IDL.Record({
     'has_created_neuron_recipes' : IDL.Opt(IDL.Bool),
+    'hotkeys' : IDL.Opt(Principals),
     'nns_neuron_id' : IDL.Nat64,
     'amount_icp_e8s' : IDL.Nat64,
   });
   const CfParticipant = IDL.Record({
+    'controller' : IDL.Opt(IDL.Principal),
     'hotkey_principal' : IDL.Text,
     'cf_neurons' : IDL.Vec(CfNeuron),
   });
@@ -1483,6 +1531,7 @@ export const init = ({ IDL }) => {
     'dissolve_state' : IDL.Opt(DissolveState),
     'followees' : IDL.Vec(IDL.Tuple(IDL.Int32, Followees)),
     'neuron_fees_e8s' : IDL.Nat64,
+    'visibility' : IDL.Opt(IDL.Int32),
     'transfer' : IDL.Opt(NeuronStakeTransfer),
     'known_neuron_data' : IDL.Opt(KnownNeuronData),
     'spawn_at_timestamp_seconds' : IDL.Opt(IDL.Nat64),