+++ packages/nns/candid/governance.d.ts
@@ -8,7 +8,9 @@ export interface AccountIdentifier {
 export type Action =
   | { RegisterKnownNeuron: KnownNeuron }
   | { ManageNeuron: ManageNeuron }
+  | { UpdateCanisterSettings: UpdateCanisterSettings }
   | { InstallCode: InstallCode }
+  | { StopOrStartCanister: StopOrStartCanister }
   | { CreateServiceNervousSystem: CreateServiceNervousSystem }
   | { ExecuteNnsFunction: ExecuteNnsFunction }
   | { RewardNodeProvider: RewardNodeProvider }
@@ -47,6 +49,14 @@ export type By =
 export interface Canister {
   id: [] | [Principal];
 }
+export interface CanisterSettings {
+  freezing_threshold: [] | [bigint];
+  controllers: [] | [Controllers];
+  log_visibility: [] | [number];
+  wasm_memory_limit: [] | [bigint];
+  memory_allocation: [] | [bigint];
+  compute_allocation: [] | [bigint];
+}
 export interface CanisterStatusResultV2 {
   status: [] | [number];
   freezing_threshold: [] | [bigint];
@@ -62,10 +72,12 @@ export interface CanisterSummary {
 }
 export interface CfNeuron {
   has_created_neuron_recipes: [] | [boolean];
+  hotkeys: [] | [Principals];
   nns_neuron_id: bigint;
   amount_icp_e8s: bigint;
 }
 export interface CfParticipant {
+  controller: [] | [Principal];
   hotkey_principal: string;
   cf_neurons: Array<CfNeuron>;
 }
@@ -135,6 +147,9 @@ export interface Committed_1 {
 }
 export interface Configure {
   operation: [] | [Operation];
+}
+export interface Controllers {
+  controllers: Array<Principal>;
 }
 export interface Countries {
   iso_codes: Array<string>;
@@ -438,6 +453,7 @@ export interface Neuron {
   dissolve_state: [] | [DissolveState];
   followees: Array<[number, Followees]>;
   neuron_fees_e8s: bigint;
+  visibility: [] | [number];
   transfer: [] | [NeuronStakeTransfer];
   known_neuron_data: [] | [KnownNeuronData];
   spawn_at_timestamp_seconds: [] | [bigint];
@@ -476,6 +492,7 @@ export interface NeuronInfo {
   stake_e8s: bigint;
   joined_community_fund_timestamp_seconds: [] | [bigint];
   retrieved_at_timestamp_seconds: bigint;
+  visibility: [] | [number];
   known_neuron_data: [] | [KnownNeuronData];
   voting_power: bigint;
   age_seconds: bigint;
@@ -578,6 +595,7 @@ export type Operation =
   | { StopDissolving: {} }
   | { StartDissolving: {} }
   | { IncreaseDissolveDelay: IncreaseDissolveDelay }
+  | { SetVisibility: SetVisibility }
   | { JoinCommunityFund: {} }
   | { LeaveCommunityFund: {} }
   | { SetDissolveTimestamp: SetDissolveTimestamp };
@@ -718,6 +736,9 @@ export interface SetSnsTokenSwapOpenTimeWindow {
   request: [] | [SetOpenTimeWindowRequest];
   swap_canister_id: [] | [Principal];
 }
+export interface SetVisibility {
+  visibility: [] | [number];
+}
 export interface SettleCommunityFundParticipation {
   result: [] | [Result_8];
   open_sns_token_swap_proposal_id: [] | [bigint];
@@ -746,6 +767,10 @@ export interface StakeMaturity {
 export interface StakeMaturityResponse {
   maturity_e8s: bigint;
   staked_maturity_e8s: bigint;
+}
+export interface StopOrStartCanister {
+  action: [] | [number];
+  canister_id: [] | [Principal];
 }
 export interface SwapBackgroundInformation {
   ledger_index_canister_summary: [] | [CanisterSummary];
@@ -796,6 +821,10 @@ export interface TimeWindow {
 }
 export interface Tokens {
   e8s: [] | [bigint];
+}
+export interface UpdateCanisterSettings {
+  canister_id: [] | [Principal];
+  settings: [] | [CanisterSettings];
 }
 export interface UpdateNodeProvider {
   reward_account: [] | [AccountIdentifier];